#110 extract_demo_content_table_creation (#125)
